Move .reader.base.Converter → new .convert module
- Make .writer.base.BaseWriter a subclass of Converter.
